+++ modules/portal/portal-ldap/src/com/liferay/portal/ldap/internal/exportimport/DefaultLDAPToPortalConverter.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.ldap.exportimport;
+package com.liferay.portal.ldap.internal.exportimport;
 
 import com.liferay.portal.UserEmailAddressException;
 import com.liferay.portal.UserScreenNameException;
