+++ modules/portal/portal-ldap/src/com/liferay/portal/ldap/internal/model/listener/ContactModelListener.java
@@ -12,10 +12,10 @@
  * details.
  */
 
-package com.liferay.portal.ldap.model.listener;
+package com.liferay.portal.ldap.internal.model.listener;
 
 import com.liferay.portal.ModelListenerException;
-import com.liferay.portal.ldap.exportimport.UserImportTransactionThreadLocal;
+import com.liferay.portal.ldap.internal.UserImportTransactionThreadLocal;
 import com.liferay.portal.model.BaseModelListener;
 import com.liferay.portal.model.Contact;
 import com.liferay.portal.model.ModelListener;
