+++ modules/portal/portal-ldap/src/com/liferay/portal/ldap/internal/model/listener/UserModelListener.java
@@ -12,12 +12,12 @@
  * details.
  */
 
-package com.liferay.portal.ldap.model.listener;
+package com.liferay.portal.ldap.internal.model.listener;
 
 import com.liferay.portal.ModelListenerException;
 import com.liferay.portal.kernel.language.LanguageUtil;
 import com.liferay.portal.kernel.util.GetterUtil;
-import com.liferay.portal.ldap.exportimport.UserImportTransactionThreadLocal;
+import com.liferay.portal.ldap.internal.UserImportTransactionThreadLocal;
 import com.liferay.portal.model.BaseModelListener;
 import com.liferay.portal.model.Group;
 import com.liferay.portal.model.MembershipRequest;
